Refactored createMarkers to separate curried function
