+++ client/src/layouts/Mapbox/mapboxInit.js
@@ -1,12 +1,13 @@
 import mapboxgl from "mapbox-gl";
 import server from "../../api/config";
 import mapPin from "./Marker/map-pin.png";
+import getPinGeoJson from "../../api/locationApi";
 
 const url = `${server}/location`;
 const defaultZoom = 3;
 const zoomThreshold = 3.5;
 
-const initMap = ({ setMap, mapContainer }) => {
+const initMap = ({ setMap, mapContainer, setGeoJsonData, addMarkers }) => {
   mapboxgl.accessToken = process.env.REACT_APP_MAPBOX_TOKEN;
 
   const map = new mapboxgl.Map({
@@ -29,16 +30,6 @@ const initMap = ({ setMap, mapContainer }) => {
   }
 
   const zoom = new Zoom(defaultZoom);
-
-  const flyToLabelAndZoom = (feature) => {
-    map.flyTo({
-      center: feature.geometry.coordinates,
-      offset: [0, -220], // [x, y] pixels
-      speed: 0.7,
-      zoom: 6,
-      essential: true,
-    });
-  };
 
   const toggleLabels = () => {
     const currentZoom = map.getZoom();
@@ -109,36 +100,10 @@ const initMap = ({ setMap, mapContainer }) => {
     });
 
     // Renders markers based on geojson data object
-    // getPinGeoJson()
-    //   .then((res) => {
-    //     setGeoJsonData(res);
-
-    //     res.features.forEach((marker) => {
-    //       const { id, city, province, country, count } = marker.properties;
-    //       const markerContainer = document.createElement("div");
-
-    //       const clickHandler = (e) => {
-    //         flyToLabelAndZoom(marker);
-    //         handleDrawer(true);
-    //         setLocation({ city, province, country });
-    //         getTributesForLocation(id).then((res) => {
-    //           setMemorials(res.data);
-    //         });
-    //       };
-
-    // Based on mapbox's implmentation, probably not the most optimal at the moment
-    // These components cannot read updated state in the Mapbox component
-    /* eslint-disable react/no-render-return-value */
-    //     const labelEl = ReactDOM.render(
-    //       <div data-location={id}>
-    //         <Label background="light" count={count} clickHandler={clickHandler} />
-    //       </div>,
-    //       markerContainer
-    //     );
-    //     new mapboxgl.Marker(labelEl).setLngLat(marker.geometry.coordinates).addTo(map);
-    //   });
-    // })
-    // .catch((err) => console.error(err));
+    getPinGeoJson().then((data) => {
+      setGeoJsonData(data);
+      addMarkers(map)(data);
+    });
 
     map.on("zoom", () => {
       toggleLabels();
